+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 # Changelog — `latexrestricted` Python package
 
 
+## v0.4.0 (dev)
+
+* Under Windows with TeX Live, the `SELFAUTOLOC` environment variable is no
+  longer used to locate `kpsewhich`.  Under these conditions, a shell escape
+  executable will often be launched with TeX Live's executable wrapper
+  `runscript.exe`.  This uses `kpathsea` internally, which overwrites
+  `SELFAUTOLOC` so that it refers to the location of the executable wrapper,
+  not the location of the TeX executable that initially invoked shell escape.
+  Under these conditions, `kpsewhich` is now located by searching path with
+  Python's `shutil.which()` for a `tlmgr` executable with accompanying
+  `kpsewhich`, and there is no way to guarantee that the correct `kpsewhich`
+  is used on systems with multiple TeX Live installations.
+
+
+
 ## v0.3.0 (2024-08-11)
 
 *  Fixed `__slots__` bug in `AnyPath`.
@@ -9,9 +24,11 @@
    for better consistency.
 
 
+
 ## v0.2.0 (2024-08-10)
 
 *  Added check for unsafe `TEXMFOUTPUT` value in TeX Live config.
+
 
 
 ## v0.1.0 (2024-07-27)